lambda_build/requirements.txt
fastapi
requests
langchain
langchain-core
langchain-community
langchain-groq
langgraph
python-dotenv
sqlalchemy
mangum
